requirements.txt
# Visanté AI Engine - Production Dependencies
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
google-genai>=1.50.0
websockets>=14.0
python-dotenv>=1.0.0,<2.0
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
chromadb>=0.4.0,<1.0
sentence-transformers>=2.0,<3
pypdf>=4.0,<6
